+++ apps/hubble/src/addon/src/store/user_data_store.rs
@@ -50,6 +50,14 @@ impl StoreDef for UserDataStoreDef {
         false
     }
 
+    fn compact_state_message_type(&self) -> u8 {
+        MessageType::None as u8
+    }
+
+    fn is_compact_state_type(&self, _message: &Message) -> bool {
+        false
+    }
+
     fn find_merge_add_conflicts(&self, _db: &RocksDB, _message: &Message) -> Result<(), HubError> {
         // No conflicts
         Ok(())
@@ -88,6 +96,13 @@ impl StoreDef for UserDataStoreDef {
         Err(HubError {
             code: "bad_request.invalid_param".to_string(),
             message: "removes not supported".to_string(),
+        })
+    }
+
+    fn make_compact_state_add_key(&self, _message: &Message) -> Result<Vec<u8>, HubError> {
+        Err(HubError {
+            code: "bad_request.invalid_param".to_string(),
+            message: "UserDataStore doesn't support compact state".to_string(),
         })
     }