+++ apps/hubble/src/hubble.ts
@@ -20,6 +20,7 @@ import {
   UserNameProof,
   validations,
 } from "@farcaster/hub-nodejs";
+import { ClientOptions as StatsDClientOptions } from "@figma/hot-shots";
 import { PeerId } from "@libp2p/interface-peer-id";
 import { peerIdFromBytes, peerIdFromString } from "@libp2p/peer-id";
 import { publicAddressesFirst } from "@libp2p/utils/address-sort";
@@ -74,7 +75,7 @@ import { CheckIncomingPortsJobScheduler } from "./storage/jobs/checkIncomingPort
 import { applyNetworkConfig, fetchNetworkConfig, NetworkConfig } from "./network/utils/networkConfig.js";
 import { UpdateNetworkConfigJobScheduler } from "./storage/jobs/updateNetworkConfigJob.js";
 import { DbSnapshotBackupJobScheduler } from "./storage/jobs/dbSnapshotBackupJob.js";
-import { statsd, StatsDInitParams } from "./utils/statsd.js";
+import { statsd } from "./utils/statsd.js";
 import {
   getDbSchemaVersion,
   LATEST_DB_SCHEMA_VERSION,
@@ -88,7 +89,7 @@ import { HttpAPIServer } from "./rpc/httpServer.js";
 import { SingleBar } from "cli-progress";
 import { exportToProtobuf } from "@libp2p/peer-id-factory";
 import OnChainEventStore from "./storage/stores/onChainEventStore.js";
-import { ensureMessageData, isMessageInDB } from "./storage/db/message.js";
+import { areMessagesInDb, ensureMessageData, isMessageInDB } from "./storage/db/message.js";
 import { getFarcasterTime, HubResult, MessageBundle } from "@farcaster/core";
 import { MerkleTrie } from "./network/sync/merkleTrie.js";
 import { DEFAULT_CATCHUP_SYNC_SNAPSHOT_MESSAGE_LIMIT } from "./defaultConfig.js";
@@ -196,7 +197,7 @@ export interface HubOptions {
   trustXForwardedForHeader?: boolean;
 
   /** StatsD parameters */
-  statsdParams?: StatsDInitParams | undefined;
+  statsdParams?: StatsDClientOptions | undefined;
 
   /** ETH mainnet RPC URL(s) */
   ethMainnetRpcUrl?: string;
@@ -1661,16 +1662,16 @@ export class Hub implements HubInterface {
     const dedupedMessages: { i: number; message: Message }[] = [];
     if (source === "gossip") {
       // Go over all the messages and see if they are in the DB. If they are, don't bother processing them
-      await Promise.all(
-        messageBundle.messages.map(async (message, i) => {
-          if (await isMessageInDB(this.rocksDB, message)) {
-            log.debug({ source }, "submitMessageBundle rejected: Message already exists");
-            allResults.set(i, err(new HubError("bad_request.duplicate", "message has already been merged")));
-          } else {
-            dedupedMessages.push({ i, message: ensureMessageData(message) });
-          }
-        }),
-      );
+      const messagesExist = await areMessagesInDb(this.rocksDB, messageBundle.messages);
+
+      for (let i = 0; i < messagesExist.length; i++) {
+        if (messagesExist[i]) {
+          log.debug({ source }, "submitMessageBundle rejected: Message already exists");
+          allResults.set(i, err(new HubError("bad_request.duplicate", "message has already been merged")));
+        } else {
+          dedupedMessages.push({ i, message: ensureMessageData(messageBundle.messages[i] as Message) });
+        }
+      }
     } else {
       dedupedMessages.push(...messageBundle.messages.map((message, i) => ({ i, message: ensureMessageData(message) })));
     }
@@ -1749,6 +1750,7 @@ export class Hub implements HubInterface {
     log.info(
       {
         hash: bytesToHexString(messageBundle.hash).unwrapOr(messageBundle.hash),
+        source,
         success,
         finalFailures: [...finalFailures],
         total: finalResults.length,