+++ apps/hubble/src/network/sync/multiPeerSyncEngine.test.ts
@@ -383,14 +383,15 @@ describe("Multi peer sync engine", () => {
     expect(await syncEngine1.trie.rootHash()).toEqual(await syncEngine2.trie.rootHash());
 
     // Now, delete the messages from engine 1
-    await engine1.getDb().clear();
+    engine1.getDb().clear();
     const allValues = await syncEngine1.trie.getAllValues(new Uint8Array());
     for (const value of allValues) {
       await syncEngine1.trie.deleteByBytes(value);
     }
 
-    // Now, engine 1 should have no messages
-    expect((await syncEngine1.trie.getTrieNodeMetadata(new Uint8Array()))?.numMessages).toEqual(0);
+    // Now, engine 1 should have no messages. Getting the metadata for the root should return
+    // undefined since the root node doesn't exist any more
+    expect(await syncEngine1.trie.getTrieNodeMetadata(new Uint8Array())).toBeUndefined();
 
     const startScore = syncEngine2.getPeerScore("engine1")?.score ?? 0;
 
@@ -538,7 +539,9 @@ describe("Multi peer sync engine", () => {
     await syncEngine2.performSync("engine1", clientForServer1);
 
     // Because do it without awaiting, we need to wait for the promise to resolve
-    await sleep(100);
+    await sleepWhile(() => syncEngine2.syncTrieQSize > 0, SLEEPWHILE_TIMEOUT);
+    await sleepWhile(() => syncEngine1.syncTrieQSize > 0, SLEEPWHILE_TIMEOUT);
+
     expect(await syncEngine1.trie.items()).toEqual(await syncEngine2.trie.items());
     expect(engine1Hash).toEqual(await syncEngine2.trie.rootHash());
   });
@@ -573,8 +576,6 @@ describe("Multi peer sync engine", () => {
     await engine1.mergeOnChainEvent(storageEvent);
     await engine2.mergeOnChainEvent(storageEvent);
 
-    const initialEngine1Count = await syncEngine1.trie.items();
-    const initialEngine2Count = await syncEngine1.trie.items();
     // We'll get 2 CastAdds
     const castAdd1 = await Factories.CastAddMessage.create({ data: { fid, network } }, { transient: { signer } });
     const castAdd2 = await Factories.CastAddMessage.create({ data: { fid, network } }, { transient: { signer } });
@@ -584,22 +585,26 @@ describe("Multi peer sync engine", () => {
     expect(await engine1.mergeMessage(castAdd1)).toBeTruthy();
 
     // CastAdd2 is added only to the sync trie, but is missing from the engine
-    await syncEngine2.trie.insert(SyncId.fromMessage(castAdd2));
+    expect(await syncEngine2.trie.insert(SyncId.fromMessage(castAdd2))).toBeTruthy();
 
     // Wait for the sync trie to be updated
-    await sleepWhile(async () => (await syncEngine2.trie.items()) !== initialEngine2Count + 2, SLEEPWHILE_TIMEOUT);
-    await sleepWhile(async () => (await syncEngine1.trie.items()) !== initialEngine1Count + 1, SLEEPWHILE_TIMEOUT);
+    await sleepWhile(() => syncEngine2.syncTrieQSize > 0, SLEEPWHILE_TIMEOUT);
+    await sleepWhile(() => syncEngine1.syncTrieQSize > 0, SLEEPWHILE_TIMEOUT);
+
+    expect(await syncEngine2.trie.items()).toEqual(3 + 2); // 2 onchain events + 2 castAdds
+    expect(await syncEngine1.trie.items()).toEqual(3 + 1); // 2 onchain events + 1 castAdd
 
     // Attempt to sync engine2 <-- engine1. Engine1 has only singerAdd
     await syncEngine2.performSync("engine1", clientForServer1);
 
     // The sync engine should realize that castAdd2 is not in it's engine, so it should be removed from the sync trie
     await sleepWhile(async () => (await syncEngine2.trie.exists(SyncId.fromMessage(castAdd2))) === true, 1000);
-
     expect(await syncEngine2.trie.exists(SyncId.fromMessage(castAdd2))).toBeFalsy();
 
     // but the castAdd1 should still be there
     expect(await syncEngine2.trie.exists(SyncId.fromMessage(castAdd1))).toBeTruthy();
+
+    expect(await syncEngine1.trie.items()).toEqual(await syncEngine2.trie.items());
   });
 
   test("recovers if messages are missing from the sync trie", async () => {