+++ apps/hubble/src/network/sync/syncEngine.ts
@@ -1304,11 +1304,13 @@ class SyncEngine extends TypedEmitter<SyncEvents> {
 
         const ourNode = await this._trie.getTrieNodeMetadata(workItem.prefix);
 
+        const start = Date.now();
         const theirNodeResult = await this.curSync.rpcClient.getSyncMetadataByPrefix(
           TrieNodePrefix.create({ prefix: workItem.prefix }),
           new Metadata(),
           rpcDeadline(),
         );
+        statsd().timing("syncengine.peer.get_syncmetadata_by_prefix_ms", Date.now() - start);
 
         if (theirNodeResult.isErr()) {
           log.warn(theirNodeResult.error, `Error fetching metadata for prefix ${workItem.prefix}`);
@@ -1553,19 +1555,11 @@ class SyncEngine extends TypedEmitter<SyncEvents> {
   }
 
   private async readDbStatsFromDb(): Promise<DbStats> {
-    let numFids = 0;
-    let numFnames = 0;
-
-    await this._db.forEachIteratorByPrefix(
+    const numFids = await this._db.countKeysAtPrefix(
       Buffer.from([RootPrefix.OnChainEvent, OnChainEventPostfix.IdRegisterByFid]),
-      () => {
-        numFids += 1;
-      },
     );
 
-    await this._db.forEachIteratorByPrefix(Buffer.from([RootPrefix.FNameUserNameProof]), () => {
-      numFnames += 1;
-    });
+    const numFnames = await this._db.countKeysAtPrefix(Buffer.from([RootPrefix.FNameUserNameProof]));
 
     return {
       numItems: await this._trie.items(),