+++ apps/hubble/src/storage/engine/index.ts
@@ -6,6 +6,7 @@ import {
   CastId,
   CastRemoveMessage,
   FarcasterNetwork,
+  getDefaultStoreLimit,
   getStoreLimits,
   hexStringToBytes,
   HubAsyncResult,
@@ -13,6 +14,7 @@ import {
   HubErrorCode,
   HubEvent,
   HubResult,
+  isLinkCompactStateMessage,
   isSignerOnChainEvent,
   isUserDataAddMessage,
   isUsernameProofMessage,
@@ -350,6 +352,7 @@ class Engine extends TypedEmitter<EngineEvents> {
       const setPostfix = typeToSetPostfix(message.data!.type);
 
       switch (setPostfix) {
+        case UserPostfix.LinkCompactStateMessage:
         case UserPostfix.LinkMessage: {
           linkMessages.push({ i, message });
           break;
@@ -465,6 +468,7 @@ class Engine extends TypedEmitter<EngineEvents> {
       }
 
       switch (setPostfix) {
+        case UserPostfix.LinkCompactStateMessage:
         case UserPostfix.LinkMessage: {
           return this._linkStore.revoke(message.value);
         }
@@ -547,8 +551,8 @@ class Engine extends TypedEmitter<EngineEvents> {
   }
 
   /** revoke message if it is not valid */
-  async validateOrRevokeMessage(message: Message): HubAsyncResult<number | undefined> {
-    const isValid = await this.validateMessage(message);
+  async validateOrRevokeMessage(message: Message, lowPriority = false): HubAsyncResult<number | undefined> {
+    const isValid = await this.validateMessage(message, lowPriority);
 
     if (isValid.isErr() && message.data) {
       if (isValid.error.errCode === "unavailable.network_failure") {
@@ -558,6 +562,7 @@ class Engine extends TypedEmitter<EngineEvents> {
       const setPostfix = typeToSetPostfix(message.data.type);
 
       switch (setPostfix) {
+        case UserPostfix.LinkCompactStateMessage:
         case UserPostfix.LinkMessage: {
           return this._linkStore.revoke(message);
         }
@@ -1033,7 +1038,7 @@ class Engine extends TypedEmitter<EngineEvents> {
     return ok(event);
   }
 
-  async validateMessage(message: Message): HubAsyncResult<Message> {
+  async validateMessage(message: Message, lowPriority = false): HubAsyncResult<Message> {
     // 1. Ensure message data is present
     if (!message || !message.data) {
       return err(new HubError("bad_request.validation_failure", "message data is missing"));
@@ -1131,7 +1136,7 @@ class Engine extends TypedEmitter<EngineEvents> {
       }
     }
 
-    // For username proof messages, make sure the name resolves to the users custody address or a connected address actually owns the ens name
+    // 6. For username proof messages, make sure the name resolves to the users custody address or a connected address actually owns the ens name
     if (isUsernameProofMessage(message) && message.data.usernameProofBody.type === UserNameType.USERNAME_TYPE_ENS_L1) {
       const result = await this.validateEnsUsernameProof(message.data.usernameProofBody, custodyAddress);
       if (result.isErr()) {
@@ -1148,12 +1153,34 @@ class Engine extends TypedEmitter<EngineEvents> {
       }
     }
 
+    // LinkCompactStateMessages can't be more than 100 storage units
+    if (
+      isLinkCompactStateMessage(message) &&
+      message.data.linkCompactStateBody.targetFids.length > getDefaultStoreLimit(StoreType.LINKS) * 100
+    ) {
+      return err(
+        new HubError("bad_request.validation_failure", "LinkCompactStateMessage is too big. Limit = 100 storage units"),
+      );
+    }
+
     // 6. Check message body and envelope
     if (this._validationWorkers) {
       this._nextValidationWorker += 1;
       this._nextValidationWorker = this._nextValidationWorker % this._validationWorkers.length;
 
-      const worker = this._validationWorkers[this._nextValidationWorker] as Worker;
+      // If this is a low-priority message and we're under load, only send it to the [0] worker,
+      // leaving the rest for high-priority messages
+      let workerIndex = this._nextValidationWorker;
+      if (this._validationWorkerPromiseMap.size > 100) {
+        if (lowPriority) {
+          workerIndex = 0;
+        } else {
+          // Send the high-priority message any but the first worker, which is reserved for the low-priority messages
+          workerIndex = this._nextValidationWorker === 0 ? 1 : this._nextValidationWorker;
+        }
+      }
+
+      const worker = this._validationWorkers[workerIndex] as Worker;
       return new Promise<HubResult<Message>>((resolve) => {
         const id = this._validationWorkerJobId++;
         this._validationWorkerPromiseMap.set(id, resolve);