+++ packages/shuttle/src/example-app/env.ts
@@ -11,6 +11,9 @@ export const HUB_SSL = process.env["HUB_SSL"] === "true" ? true : false;
 export const POSTGRES_URL = process.env["POSTGRES_URL"] || "postgres://localhost:5432";
 export const REDIS_URL = process.env["REDIS_URL"] || "redis://localhost:6379";
 
+export const TOTAL_SHARDS = parseInt(process.env["SHARDS"] || "0");
+export const SHARD_INDEX = parseInt(process.env["SHARD_NUM"] || "0");
+
 export const BACKFILL_FIDS = process.env["FIDS"] || "";
 export const MAX_FID = process.env["MAX_FID"];